PR-D first slice: extract api/_core + api/auth, barrel-re-export from lib/api
Splits the monolithic 2711-line src/lib/api.ts into a barrel that
re-exports from domain files under src/lib/api/. First two extractions:

- src/lib/api/_core.ts (159 lines): API, BACKEND_ROOT, ProviderKeyMissingError,
  ProviderUpstreamError + parsers, jsonOrThrow, apiFetch.
- src/lib/api/auth.ts (124 lines): AUTH, CurrentUser, AuthError, getCurrentUser,
  signin/signout/signup/forgotPassword/resetPassword/verifyEmail/requestVerifyToken,
  UserProfileUpdate, updateProfile.

src/lib/api.ts shrinks 2711 -> 2484 lines; barrel re-exports every public
symbol via export {...} from "./api/_core" / "./api/auth" so existing
`../lib/api` import paths keep resolving everywhere. No call-site changes
required.

Pattern is now in place for the remaining domain extractions (matters,
documents, signoffs, modules, reviews, audit, ceremony, lawve) as
mechanical follow-ups.

Verification:
- frontend typecheck clean
- frontend vitest: 193/193 passed
- frontend prod build: green
